+++ src/formElementPlugins/listBuilder/src/displayComponents/ListCardLayout.tsx
@@ -1,0 +1,50 @@
+import React from 'react'
+import { Card, Icon, Label } from 'semantic-ui-react'
+import { ListLayoutProps } from '../types'
+import { substituteValues } from '../helpers'
+
+const ListCardLayout: React.FC<ListLayoutProps> = ({
+  listItems,
+  displayFormat,
+  editItem = () => {},
+  deleteItem = () => {},
+  Markdown,
+  isEditable = true,
+}) => {
+  const { title, subtitle, description } = displayFormat
+  return (
+    <>
+      {listItems.map((item, index) => (
+        <Card key={`list-item-${index}`}>
+          <Card.Content>
+            {isEditable && (
+              <Label floating onClick={() => deleteItem(index)}>
+                <Icon name="delete" />
+              </Label>
+            )}
+            {title && (
+              <Card.Header onClick={() => editItem(index)}>
+                <Markdown text={substituteValues(title, item)} semanticComponent="noParagraph" />
+              </Card.Header>
+            )}
+            {subtitle && (
+              <Card.Meta onClick={() => editItem(index)}>
+                <Markdown text={substituteValues(subtitle, item)} semanticComponent="noParagraph" />
+              </Card.Meta>
+            )}
+            {description && (
+              <Card.Description onClick={() => editItem(index)}>
+                <Markdown
+                  text={substituteValues(description, item)}
+                  semanticComponent="noParagraph"
+                />
+              </Card.Description>
+            )}
+          </Card.Content>
+        </Card>
+      ))}
+    </>
+  )
+}
+
+export default ListCardLayout
